+++ crates/zng/src/lib.rs
@@ -559,7 +559,6 @@ pub mod prelude {
     pub use crate::__prelude::*;
 }
 mod __prelude {
-    // TODO(breaking) add FrequencyUnits
     pub use crate::{color, gesture, keyboard, layout, mouse, task, timer, touch, widget};
 
     pub use zng_task::rayon::prelude::{
@@ -594,8 +593,8 @@ mod __prelude {
     pub use crate::var::animation::easing;
 
     pub use zng_layout::unit::{
-        Align, AngleUnits as _, ByteUnits as _, DipToPx as _, FactorUnits as _, Layout1d as _, Layout2d as _, Length, LengthUnits as _,
-        LineFromTuplesBuilder as _, PxDensityUnits as _, PxToDip as _, RectFromTuplesBuilder as _, TimeUnits as _,
+        Align, AngleUnits as _, ByteUnits as _, DipToPx as _, FactorUnits as _, FrequencyUnits as _, Layout1d as _, Layout2d as _, Length,
+        LengthUnits as _, LineFromTuplesBuilder as _, PxDensityUnits as _, PxToDip as _, RectFromTuplesBuilder as _, TimeUnits as _,
     };
 
     pub use zng_txt::{ToTxt as _, Txt, formatx};
@@ -803,10 +802,10 @@ mod __prelude_wgt {
         context::{DIRECTION_VAR, LAYOUT, LayoutDirection, LayoutMetrics},
         unit::{
             Align, AngleDegree, AngleGradian, AngleRadian, AngleUnits as _, ByteUnits as _, Dip, DipBox, DipPoint, DipRect, DipSideOffsets,
-            DipSize, DipToPx as _, DipVector, Factor, Factor2d, FactorPercent, FactorSideOffsets, FactorUnits as _, Layout1d as _,
-            Layout2d as _, LayoutAxis, Length, LengthUnits as _, Line, LineFromTuplesBuilder as _, Point, Px, PxBox, PxConstraints,
-            PxConstraints2d, PxCornerRadius, PxDensityUnits as _, PxLine, PxPoint, PxRect, PxSideOffsets, PxSize, PxToDip as _,
-            PxTransform, PxVector, Rect, RectFromTuplesBuilder as _, SideOffsets, Size, TimeUnits as _, Transform, Vector,
+            DipSize, DipToPx as _, DipVector, Factor, Factor2d, FactorPercent, FactorSideOffsets, FactorUnits as _, FrequencyUnits as _,
+            Layout1d as _, Layout2d as _, LayoutAxis, Length, LengthUnits as _, Line, LineFromTuplesBuilder as _, Point, Px, PxBox,
+            PxConstraints, PxConstraints2d, PxCornerRadius, PxDensityUnits as _, PxLine, PxPoint, PxRect, PxSideOffsets, PxSize,
+            PxToDip as _, PxTransform, PxVector, Rect, RectFromTuplesBuilder as _, SideOffsets, Size, TimeUnits as _, Transform, Vector,
         },
     };